+++ CHANGELOG.md
@@ -13,6 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![API] Add missing properties methods in extensions points [#781](https://github.com/nbbrd/sdmx-dl/issues/781)
 - ![API] Add support of catalogs
 - ![CLI] Add check sources command [#515](https://github.com/nbbrd/sdmx-dl/issues/515)
+- ![GRPC] Add error mapping in rest service
 
 ### Changed